Initial commit
Currently have working dashboard for daily recommendations given a portfolio of tickers to consider, and backtest which seems to be working pretty well other than the returns stats for buy&hold which seem to be correct for an individual stock but incorrectly normalized or tracked for a portfolio. Next major step will be to implement optimization of strategy settings.